+++ ICD.Connect.Calendaring.Google/Properties/AssemblyInfo.cs
@@ -4,4 +4,4 @@
 [assembly: AssemblyCompany("ICD Systems")]
 [assembly: AssemblyProduct("ICD.Connect.Calendaring.Google")]
 [assembly: AssemblyCopyright("Copyright © ICD Systems 2021")]
-[assembly: AssemblyVersion("9.3.0.0")]
+[assembly: AssemblyVersion("9.4.0.0")]