+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [9.4.0] - 2021-05-14
 ### Added
  - Added event to CalendarManager for when the current booking changes.
  - Expose AsureDevice configured URI info as a new console node.